backend/requirements.txt
google-generativeai==0.3.2
Flask==3.0.0
Pillow>=10.4.0
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
firebase-admin==6.3.0
google-cloud-aiplatform>=1.38.0
numpy>=1.26.0
python-dotenv>=1.0.0
